requirements-tests.txt
-e .
coverage[toml]>=6.5
pytest-cov>=4.1.0
